+++ AvgScript/src/webview/_mdToHtml.ts
@@ -1,0 +1,5 @@
+import marked = require('marked');
+
+export function markdownParser(markdown: string) {
+    return marked.marked.parse(markdown);
+}